requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.2
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.7.0
